cambio nome file + cancellazione file

+++ admin/cancella.php
@@ -1,8 +1,8 @@
 <?php
 
 include_once '../inc/db.config.php';
+include_once '../inc/functions.php';
 
-session_start();
 if($_SESSION['login'] == false)
 {
     header('location: ../index.php');
@@ -10,6 +10,20 @@ if($_SESSION['login'] == false)
 }
 
 $id = $_GET['idpratica'];
+
+$sql = "SELECT documenti FROM pratiche WHERE id_pratica=?";
+
+$stmt = $conn -> prepare($sql);
+$stmt->bind_param("i", $id);
+
+if ($stmt -> execute() === FALSE ) {
+    die('NON POSSO LEGGERE LE PRATICHE NEL DATABASE' . $stmt -> error);
+};
+
+$risultati = $stmt -> get_result();
+$riga = $risultati -> fetch_assoc();
+
+elimina_file('../' . $riga['documenti']);
 
 $sql = "DELETE FROM pratiche WHERE id_pratica=?";
